Added support for multiple idea factories
Using pre-existing or specially created categories it is possible to specify an idea_category to use in the inline code added to the Wordpress page. In the below example, the category to be used for the voting is “IdeaFactory1”:

[idea_factory hide_submit="off" hide_votes="off" hide_voting="off" idea_category=“IdeaFactory1”]

By leaving out the idea_category, all ideas will be shown.